// frontend/app/lib/demo-data/types.ts
export type DashboardRole = "client" | "user" | "admin" | "superadmin";

export interface UserCreate {
  full_name: string;
  username?: string;
  email: string;
  phone?: string | null;
  password: string;
  role?: DashboardRole;
  address?: string | null;
  gouvernorat_id?: number | null;
}

export interface UserLogin {
  identifier?: string;
  email?: string;
  password: string;
}

export interface UserResponse {
  _id?: string | null;
  id: string;
  email: string;
  role?: DashboardRole;
  full_name?: string | null;
  username?: string | null;
  picture?: string | null;
  address?: string | null;
  birthdate?: string | null;
  phone?: string | null;
  gouvernorat_id?: number | null;
  is_verified?: boolean;
  is_banned?: boolean;
  created_at?: string | null;
}